move fix to shouldDisplayViolationsRBRInLHN
